refactor: Liquor config 분리
