+++ src/main/java/com/woowacamp/soolsool/core/liquor/config/LiquorConfig.java
@@ -1,14 +1,11 @@
-package com.woowacamp.soolsool.global.config;
+package com.woowacamp.soolsool.core.liquor.config;
 
-import com.woowacamp.soolsool.core.liquor.controller.converter.LiquorBrewTypeConverter;
-import com.woowacamp.soolsool.core.liquor.controller.converter.LiquorRegionTypeConverter;
-import com.woowacamp.soolsool.core.liquor.controller.converter.LiquorStatusTypeConverter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-public class WebConfig implements WebMvcConfigurer {
+public class LiquorConfig implements WebMvcConfigurer {
 
     @Override
     public void addFormatters(final FormatterRegistry registry) {
