+++ src/main/java/com/woowacamp/soolsool/core/liquor/config/LiquorRegionTypeConverter.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.liquor.controller.converter;
+package com.woowacamp.soolsool.core.liquor.config;
 
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorRegionType;
 import java.util.Arrays;
@@ -11,8 +11,8 @@ public class LiquorRegionTypeConverter implements Converter<String, LiquorRegion
     @Override
     public LiquorRegionType convert(final String regionType) {
         return Arrays.stream(LiquorRegionType.values())
-            .filter(value -> Objects.equals(value.getName(), regionType))
-            .findFirst()
-            .orElse(null);
+                .filter(value -> Objects.equals(value.getName(), regionType))
+                .findFirst()
+                .orElse(null);
     }
 }
