+++ src/main/java/com/woowacamp/soolsool/core/liquor/config/LiquorStatusTypeConverter.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.liquor.controller.converter;
+package com.woowacamp.soolsool.core.liquor.config;
 
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorStatusType;
 import java.util.Arrays;
@@ -11,8 +11,8 @@ public class LiquorStatusTypeConverter implements Converter<String, LiquorStatus
     @Override
     public LiquorStatusType convert(final String statusType) {
         return Arrays.stream(LiquorStatusType.values())
-            .filter(value -> Objects.equals(value.getStatus(), statusType))
-            .findFirst()
-            .orElse(null);
+                .filter(value -> Objects.equals(value.getStatus(), statusType))
+                .findFirst()
+                .orElse(null);
     }
 }
